+++ modules/json/src/bundled.ts
@@ -8,8 +8,8 @@ export type {JSONTableLoaderOptions} from './json-table-loader-with-parser';
 export {JSONTableLoaderWithParser as JSONTableLoader} from './json-table-loader-with-parser';
 export type {NDJSONLoaderOptions} from './ndjson-loader-with-parser';
 export {NDJSONLoaderWithParser as NDJSONLoader} from './ndjson-loader-with-parser';
-export type {NDGeoJSONLoaderOptions} from './ndgeoson-loader-with-parser';
-export {NDJSONLoaderWithParser as NDGeoJSONLoader} from './ndgeoson-loader-with-parser';
+export type {NDGeoJSONLoaderOptions} from './ndgeojson-loader-types';
+export {NDGeoJSONLoaderWithParser as NDGeoJSONLoader} from './ndgeojson-loader';
 
 export type {GeoJSONLoaderOptions} from './geojson-loader-with-parser';
 export {